refactor: home bias, pool, rate-limit, and jobs tests
Move remaining large inline test modules next to their impls and
split the 800-line jobs dump into tests/{store,queue,events}.rs.
